requirements.txt
asn1crypto==0.24.0
astroid==1.6.5
backports.functools-lru-cache==1.5
configparser==3.5.0
cryptography==2.1.4
Django==1.11.18
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==3.3
enum34==1.1.6
future==0.17.1
futures==3.2.0
idna==2.6
ipaddress==1.0.17
isort==4.3.4
keyring==10.6.0
keyrings.alt==3.0
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
Pillow==5.4.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.7.1
pylint==1.9.4
pytz==2018.9
pyxdg==0.25
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.12.0
virtualenv==16.3.0
wrapt==1.10.11